+++ packages/core/styles/src/packages/atomic/icon/icon.style.css.ts
@@ -3,4 +3,5 @@ import { style as vanillaStyle } from "@vanilla-extract/css";
 export const container = vanillaStyle({
   display: "flex",
   alignItems: "center",
+  pointerEvents: "all",
 });