+++ packages/core/styles/src/packages/atomic/title/title.sprinkle.css.ts
@@ -1,6 +1,13 @@
 import { defineProperties, createSprinkles } from "@vanilla-extract/sprinkles";
 
-import { colorProperties, textAlignProperties } from "../../../properties";
+import {
+  colorProperties,
+  titleFontSizeProperties,
+  fontWeightProperties,
+  titleLineHeightProperties,
+  textAlignProperties,
+  sizeProperties,
+} from "../../../properties";
 
 const titleColorProperties = {
   "primary.textHigh": colorProperties["primary.textHigh"],
@@ -15,9 +22,22 @@ const titleColorProperties = {
   "neutral.textLow": colorProperties["neutral.textLow"],
 };
 
+const fontSizeProperties = {
+  ...titleFontSizeProperties,
+  ...sizeProperties,
+};
+
+const lineHeightProperties = {
+  ...titleLineHeightProperties,
+  ...sizeProperties,
+};
+
 const properties = {
   color: titleColorProperties,
   textAlign: textAlignProperties,
+  fontSize: fontSizeProperties,
+  fontWeight: fontWeightProperties,
+  lineHeight: lineHeightProperties,
 };
 
 const sprinkle = createSprinkles(