+++ packages/core/styles/src/themes/globals.css.ts
@@ -152,6 +152,7 @@ createGlobalTheme(":root", vars, {
     xs: breakpoint.xs.value,
     md: breakpoint.md.value,
     lg: breakpoint.lg.value,
+    xl: breakpoint.xl.value,
   },
   utils: {
     focus: `0 0 0 3px ${colors.primary["surface-highlight"].value as string}`,