+++ packages/core/styles/src/themes/mediaQueries.ts
@@ -6,4 +6,5 @@ export const mediaQueries = {
   xs: () => `screen and (min-width: ${breakpoint.xs.value})`,
   md: () => `screen and (min-width: ${breakpoint.md.value})`,
   lg: () => `screen and (min-width: ${breakpoint.lg.value})`,
+  xl: () => `screen and (min-width: ${breakpoint.xl.value})`,
 };