+++ packages/react/src/atomic/Title/src/title.spec.tsx
@@ -19,42 +19,42 @@ describe("GIVEN <Title />", () => {
       ).toBeDefined();
     });
 
-    it('THEN should render element type="h1"', () => {
+    it("THEN should render element type='h1'", () => {
       makeSut({ as: "h1" });
       expect(
         screen.getByRole("heading", { name: content, level: 1 })
       ).toBeDefined();
     });
 
-    it('THEN should render element type="h2"', () => {
+    it("THEN should render element type='h2'", () => {
       makeSut({ as: "h2" });
       expect(
         screen.getByRole("heading", { name: content, level: 2 })
       ).toBeDefined();
     });
 
-    it('THEN should render element type="h3"', () => {
+    it("THEN should render element type='h3'", () => {
       makeSut({ as: "h3" });
       expect(
         screen.getByRole("heading", { name: content, level: 3 })
       ).toBeDefined();
     });
 
-    it('THEN should render element type="h4"', () => {
+    it("THEN should render element type='h4'", () => {
       makeSut({ as: "h4" });
       expect(
         screen.getByRole("heading", { name: content, level: 4 })
       ).toBeDefined();
     });
 
-    it('THEN should render element type="h5"', () => {
+    it("THEN should render element type='h5'", () => {
       makeSut({ as: "h5" });
       expect(
         screen.getByRole("heading", { name: content, level: 5 })
       ).toBeDefined();
     });
 
-    it('THEN should render element type="h6"', () => {
+    it("THEN should render element type='h6'", () => {
       makeSut({ as: "h6" });
       expect(
         screen.getByRole("heading", { name: content, level: 6 })
@@ -97,6 +97,117 @@ describe("GIVEN <Title />", () => {
           .getByRole("heading", { name: content, level: 1 })
           .getAttribute("class")
       ).toContain("textAlign_justify");
+    });
+  });
+
+  describe("THEN should correctly render a fontWeight", () => {
+    it("THEN should correctly render a bold", () => {
+      makeSut({ fontWeight: "bold" });
+      expect(screen.getByRole("heading").getAttribute("class")).toContain(
+        "fontWeight_bold"
+      );
+    });
+
+    it("THEN should correctly render a medium", () => {
+      makeSut({ fontWeight: "medium" });
+      expect(screen.getByRole("heading").getAttribute("class")).toContain(
+        "fontWeight_medium"
+      );
+    });
+
+    it("THEN should correctly render a regular", () => {
+      makeSut({ fontWeight: "regular" });
+      expect(screen.getByRole("heading").getAttribute("class")).toContain(
+        "fontWeight_regular"
+      );
+    });
+  });
+
+  describe("THEN should correctly render the size", () => {
+    it("THEN should correctly render the size h1", () => {
+      makeSut({ fontSize: "h1" });
+      expect(screen.getByRole("heading").getAttribute("class")).toContain(
+        "fontSize_h1"
+      );
+    });
+
+    it("THEN should correctly render the size h2", () => {
+      makeSut({ fontSize: "h2" });
+      expect(screen.getByRole("heading").getAttribute("class")).toContain(
+        "fontSize_h2"
+      );
+    });
+
+    it("THEN should correctly render the size h3", () => {
+      makeSut({ fontSize: "h3" });
+      expect(screen.getByRole("heading").getAttribute("class")).toContain(
+        "fontSize_h3"
+      );
+    });
+
+    it("THEN should correctly render the size h4", () => {
+      makeSut({ fontSize: "h4" });
+      expect(screen.getByRole("heading").getAttribute("class")).toContain(
+        "fontSize_h4"
+      );
+    });
+
+    it("THEN should correctly render the size h5", () => {
+      makeSut({ fontSize: "h5" });
+      expect(screen.getByRole("heading").getAttribute("class")).toContain(
+        "fontSize_h5"
+      );
+    });
+
+    it("THEN should correctly render the size h6", () => {
+      makeSut({ fontSize: "h6" });
+      expect(screen.getByRole("heading").getAttribute("class")).toContain(
+        "fontSize_h6"
+      );
+    });
+  });
+
+  describe("THEN should correctly render the line height", () => {
+    it("THEN should correctly render the line height h1", () => {
+      makeSut({ lineHeight: "h1" });
+      expect(screen.getByRole("heading").getAttribute("class")).toContain(
+        "lineHeight_h1"
+      );
+    });
+
+    it("THEN should correctly render the line height h2", () => {
+      makeSut({ lineHeight: "h2" });
+      expect(screen.getByRole("heading").getAttribute("class")).toContain(
+        "lineHeight_h2"
+      );
+    });
+
+    it("THEN should correctly render the line height h3", () => {
+      makeSut({ lineHeight: "h3" });
+      expect(screen.getByRole("heading").getAttribute("class")).toContain(
+        "lineHeight_h3"
+      );
+    });
+
+    it("THEN should correctly render the line height h4", () => {
+      makeSut({ lineHeight: "h4" });
+      expect(screen.getByRole("heading").getAttribute("class")).toContain(
+        "lineHeight_h4"
+      );
+    });
+
+    it("THEN should correctly render the line height h5", () => {
+      makeSut({ lineHeight: "h5" });
+      expect(screen.getByRole("heading").getAttribute("class")).toContain(
+        "lineHeight_h5"
+      );
+    });
+
+    it("THEN should correctly render the line height h6", () => {
+      makeSut({ lineHeight: "h6" });
+      expect(screen.getByRole("heading").getAttribute("class")).toContain(
+        "lineHeight_h6"
+      );
     });
   });
 });