+++ packages/core/styles/CHANGELOG.md
@@ -2,6 +2,22 @@
 
 Nimbus Styles deprive all styles needed to build nimbus components.
 
+## 2023-02-23 `6.12.0`
+
+#### 🎉 New features
+
+- Added `table-layout: fixed` css property to `Table` styles. ([#106](https://github.com/TiendaNube/nimbus-design-system/pull/106) by [@juanchigallego](https://github.com/juanchigallego))
+- Added `zIndex` property to `Box` sprinkles. ([#106](https://github.com/TiendaNube/nimbus-design-system/pull/106) by [@juanchigallego](https://github.com/juanchigallego))
+- Added support for new `xl` breakpoint on conditions. ([#106](https://github.com/TiendaNube/nimbus-design-system/pull/106) by [@juanchigallego](https://github.com/juanchigallego))
+- Added `fontWeight`, `fontSize` and `lineHeight` properties to `Title` component sprinkles. ([#106](https://github.com/TiendaNube/nimbus-design-system/pull/106) by [@juanchigallego](https://github.com/juanchigallego))
+- Added `sizeProperties` tokens to `Title` component sprinkles. ([#106](https://github.com/TiendaNube/nimbus-design-system/pull/106) by [@juanchigallego](https://github.com/juanchigallego))
+- Added `inherit` to `cursor` property values. ([#106](https://github.com/TiendaNube/nimbus-design-system/pull/106) by [@juanchigallego](https://github.com/juanchigallego))
+
+#### 🐛 Bug fixes
+
+- Added 100% width on `Input` container. ([#106](https://github.com/TiendaNube/nimbus-design-system/pull/106) by [@juanchigallego](https://github.com/juanchigallego))
+- Added `overflowY: auto` on `Sidebar` container. ([#106](https://github.com/TiendaNube/nimbus-design-system/pull/106) by [@juanchigallego](https://github.com/juanchigallego))
+
 ## 2023-02-15 `6.11.0`
 
 #### 🎉 New features