+++ packages/react/src/atomic/Title/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 Tilte is a basic component that allows you to give titles and more hierarchy to blocks of text for sections or header components.
 
+## 2023-02-23 `2.1.0`
+
+- Added `fontSize`, `fontWeight` and `lineHeight` properties to component API. ([#106](https://github.com/TiendaNube/nimbus-design-system/pull/106) by [@juanchigallego](https://github.com/juanchigallego))
+
 ## 2022-12-22 `2.0.0`
 
 ### 💡 Others