+++ packages/react/src/composite/Table/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 Tables allow users to display large amounts of data in a tabular manner, using columns and rows to group the different types of data.
 
+## 2023-01-15 `1.1.1`
+
+### 🐛 Bug fixes
+
+- Fixed a bug where `rest` condition selector sprinkle would not work. ([#106](https://github.com/TiendaNube/nimbus-design-system/pull/106) by [@juanchigallego](https://github.com/juanchigallego))
+
 ## 2023-01-15 `1.1.0`
 
 ### 🎉 New features